Single-Sign-On/requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
Django~=3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-phonenumber-field==5.0.0
djangorestframework~=3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
Markdown==3.2.2
phonenumbers==8.12.9
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.2.0
six==1.15.0
django-utils-six==2.0
